+++ 2686V/Source/OPN/OpnOperator.h
@@ -8,5 +8,5 @@ public:
 	OpnOperator() : FmOperator() {}
 
 	void noteOn(float frequency, float velocity, int noteNumber) override;
-	void getSample(float& output, float modulator, float amLfoVal, float pmLfoVal, bool globalPm, bool globalAm, int globalPms, int globalAms, float globalPmd = -1.0f, float globalAmd = -1.0f, float modWheel = 0.0f) override;
+	void getSample(float& output, float modulator, float amLfoVal, float pmLfoVal, bool globalPm, bool globalAm, float globalPms, float globalAms, float globalPmd = -1.0f, float globalAmd = -1.0f, float modWheel = 0.0f) override;
 };
